Move env tests to one file.
Fixes issue with importing a TestCase-like
class in another unittest file.
